Move HTTP debug code to pkg
