+++ cmd/ctr/commands/resolver.go
@@ -23,10 +23,7 @@ import (
 	"crypto/x509"
 	"errors"
 	"fmt"
-	"io"
 	"net/http"
-	"net/http/httptrace"
-	"net/http/httputil"
 	"os"
 	"strings"
 
@@ -35,6 +32,7 @@ import (
 	"github.com/containerd/containerd/v2/core/remotes/docker"
 	"github.com/containerd/containerd/v2/core/remotes/docker/config"
 	"github.com/containerd/containerd/v2/core/transfer/registry"
+	"github.com/containerd/containerd/v2/pkg/httpdbg"
 	"github.com/containerd/log"
 	"github.com/urfave/cli/v2"
 )
@@ -104,9 +102,9 @@ func GetResolver(ctx context.Context, cliContext *cli.Context) (remotes.Resolver
 
 	if cliContext.Bool("http-dump") {
 		hostOptions.UpdateClient = func(client *http.Client) error {
-			client.Transport = &DebugTransport{
-				transport: client.Transport,
-				writer:    log.G(ctx).Writer(),
+			client.Transport = &httpdbg.DebugTransport{
+				Transport: client.Transport,
+				Writer:    log.G(ctx).Writer(),
 			}
 			return nil
 		}
@@ -155,65 +153,6 @@ func resolverDefaultTLS(cliContext *cli.Context) (*tls.Config, error) {
 	}
 
 	return tlsConfig, nil
-}
-
-// DebugTransport wraps the underlying http.RoundTripper interface and dumps all requests/responses to the writer.
-type DebugTransport struct {
-	transport http.RoundTripper
-	writer    io.Writer
-}
-
-// RoundTrip dumps request/responses and executes the request using the underlying transport.
-func (t DebugTransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	in, err := httputil.DumpRequestOut(req, true)
-	if err != nil {
-		return nil, fmt.Errorf("failed to dump request: %w", err)
-	}
-
-	if _, err := t.writer.Write(in); err != nil {
-		return nil, err
-	}
-
-	resp, err := t.transport.RoundTrip(req)
-	if err != nil {
-		return nil, err
-	}
-
-	out, err := httputil.DumpResponse(resp, true)
-	if err != nil {
-		return nil, fmt.Errorf("failed to dump response: %w", err)
-	}
-
-	if _, err := t.writer.Write(out); err != nil {
-		return nil, err
-	}
-
-	return resp, err
-}
-
-// NewDebugClientTrace returns a Go http trace client predefined to write DNS and connection
-// information to the log. This is used via the --http-trace flag on push and pull operations in ctr.
-func NewDebugClientTrace(ctx context.Context) *httptrace.ClientTrace {
-	return &httptrace.ClientTrace{
-		DNSStart: func(dnsInfo httptrace.DNSStartInfo) {
-			log.G(ctx).WithField("host", dnsInfo.Host).Debugf("DNS lookup")
-		},
-		DNSDone: func(dnsInfo httptrace.DNSDoneInfo) {
-			if dnsInfo.Err != nil {
-				log.G(ctx).WithField("lookup_err", dnsInfo.Err).Debugf("DNS lookup error")
-			} else {
-				log.G(ctx).WithField("result", dnsInfo.Addrs[0].String()).WithField("coalesced", dnsInfo.Coalesced).Debugf("DNS lookup complete")
-			}
-		},
-		GotConn: func(connInfo httptrace.GotConnInfo) {
-			remoteAddr := "<nil>"
-			if addr := connInfo.Conn.RemoteAddr(); addr != nil {
-				remoteAddr = addr.String()
-			}
-
-			log.G(ctx).WithField("reused", connInfo.Reused).WithField("remote_addr", remoteAddr).Debugf("Connection successful")
-		},
-	}
 }
 
 type staticCredentials struct {
